Partial fix for problems where code generated for large models was failing.  Basically, we were returning stack pointers and then using them.  More work is necessary to fix the remaining tests.

+++ ptolemy/cg/adapter/generic/program/procedural/c/adapters/ptolemy/actor/lib/conversions/Round.c
@@ -10,7 +10,7 @@ $put(output, ceil($get(input)));
 /**/
 
 /*** floorBlock() ***/
-$put(output, floor($get(input)));
+$put(output, (double)(floor(($get(input)))));
 /**/
 
 /*** roundBlock() ***/
